chore: move entrypoint to /cmd/ayd directory
