+++ .gitignore
@@ -1,2 +1,2 @@
-ayd
+/ayd
 ayd.log
