+++ Makefile
@@ -4,7 +4,7 @@ COMMIT = $(shell git rev-parse --short $(shell git describe))
 
 
 ayd: ${SOURCES}
-	CGO_ENABLED=0 go build -ldflags="-s -w -X main.version=${VERSION} -X main.commit=${COMMIT}" -trimpath .
+	CGO_ENABLED=0 go build -ldflags="-s -w -X main.version=${VERSION} -X main.commit=${COMMIT}" -trimpath -o ayd ./cmd/ayd
 
 
 .PHONY: test cover fmt resources clean install
@@ -18,13 +18,13 @@ cover:
 fmt:
 	gofmt -s -w ${SOURCES}
 
-resources: resource_windows_386.syso resource_windows_amd64.syso resource_windows_arm.syso resource_windows_arm64.syso
+resources: cmd/ayd/resource_windows_386.syso cmd/ayd/resource_windows_amd64.syso cmd/ayd/resource_windows_arm.syso cmd/ayd/resource_windows_arm64.syso
 
-%.syso: versioninfo.json
-	go run github.com/josephspurrier/goversioninfo/cmd/goversioninfo@latest -platform-specific
+%.syso: cmd/ayd/versioninfo.json
+	cd cmd/ayd && go run github.com/josephspurrier/goversioninfo/cmd/goversioninfo@latest -platform-specific
 
 clean:
-	-rm ayd ayd.log ayd_debug.log
+	-rm ayd ayd.log
 
 install: ayd
 	sudo install ./ayd /usr/local/bin/ayd
