+++ cmd/ayd/conv.go
