+++ cmd/ayd/debug.go
