+++ cmd/ayd/help.txt
