+++ cmd/ayd/oneshot.go
