+++ cmd/ayd/oneshot_test.go
