+++ cmd/ayd/testdata/log.ltsv
