+++ cmd/ayd/versioninfo.json
@@ -38,6 +38,6 @@
             "CharsetID": "04B0"
         }
     },
-    "IconPath": "internal/endpoint/static/favicon.ico",
+    "IconPath": "../../internal/endpoint/static/favicon.ico",
     "ManifestPath": ""
 }
