+++ internal/testutil/data.go
@@ -1,0 +1,8 @@
+package testutil
+
+import (
+	_ "embed"
+)
+
+//go:embed testdata/test.log
+var DummyLog string
