+++ internal/testutil/store.go
@@ -30,13 +30,10 @@ func NewStore(t testing.TB) *store.Store {
 	return NewStoreWithConsole(t, io.Discard)
 }
 
-//go:embed testdata/test.log
-var rawLog []byte
-
 func NewStoreWithLog(t testing.TB) *store.Store {
 	fpath := filepath.Join(t.TempDir(), "ayd.log")
 
-	if err := os.WriteFile(fpath, rawLog, 0644); err != nil {
+	if err := os.WriteFile(fpath, []byte(DummyLog), 0644); err != nil {
 		t.Fatalf("failed to prepare test log file: %s", err)
 	}
